Refactor: image controller를 따로 구현함에 따라 기존의 코드들 변경
