+++ src/main/java/com/zero/triptalk/application/PlannerApplication.java
@@ -2,7 +2,7 @@ package com.zero.triptalk.application;
 
 import com.zero.triptalk.exception.custom.PlannerDetailException;
 import com.zero.triptalk.place.entity.Place;
-import com.zero.triptalk.place.service.ImageService;
+import com.zero.triptalk.image.service.ImageService;
 import com.zero.triptalk.place.service.PlaceService;
 import com.zero.triptalk.planner.dto.PlannerDetailListRequest;
 import com.zero.triptalk.planner.dto.PlannerDetailRequest;
