+++ src/main/java/com/zero/triptalk/planner/service/PlannerDetailService.java
@@ -2,7 +2,7 @@ package com.zero.triptalk.planner.service;
 
 import com.zero.triptalk.exception.custom.PlannerDetailException;
 import com.zero.triptalk.exception.custom.UserException;
-import com.zero.triptalk.place.service.ImageService;
+import com.zero.triptalk.image.service.ImageService;
 import com.zero.triptalk.place.service.PlaceService;
 import com.zero.triptalk.planner.dto.PlannerDetailListResponse;
 import com.zero.triptalk.planner.dto.PlannerDetailRequest;
